Moved cmap test into their own file
