+++ test/cmap-test.h
@@ -1,0 +1,6 @@
+
+#include <cmap.h>
+#include <hash.h>
+#include <permutations.h>
+
+bool cmap_test();
